requirements.txt
Keras==2.1.6
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.3.0.38
Pillow==8.1.0
scikit-learn
tensorflow==1.5

